// cgocore-ui/src/app/features/models/reference/localization/views/full/CurrencyFullView.ts
import { CurrencyForm } from 'src/app/features/models/reference/localization/forms/CurrencyForm';
/**
 * auto generated full veiw bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class CurrencyFullView {

id: number;
form: CurrencyForm;
canUpdate: boolean;
canDelete: boolean;

/* Specific Code Start */
/* Specific Code End */
}